% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparci.R
\name{ggparci}
\alias{ggparci}
\title{FUNCTION_TITLE}
\usage{
ggparci(x, group, level = 0.95, lab_group = "", lab_y = "", lab_x = "",
  title = "", ylim, include_lines = FALSE, seriate = TRUE,
  facet = FALSE, flip = FALSE, alpha_per_line = 0.1,
  alpha_geom_ribbon = 0.5, jitter_median_factor = 0, ...)
}
\arguments{
\item{x}{should be a data.frame with each column representing a data dimension}

\item{group}{a grouping variable}

\item{level}{PARAM_DESCRIPTION, Default: 0.95}

\item{lab_group}{PARAM_DESCRIPTION, Default: ''}

\item{lab_y}{PARAM_DESCRIPTION, Default: ''}

\item{lab_x}{PARAM_DESCRIPTION, Default: ''}

\item{title}{PARAM_DESCRIPTION, Default: ''}

\item{ylim}{PARAM_DESCRIPTION}

\item{include_lines}{PARAM_DESCRIPTION, Default: FALSE}

\item{seriate}{PARAM_DESCRIPTION, Default: TRUE}

\item{facet}{PARAM_DESCRIPTION, Default: FALSE}

\item{flip}{PARAM_DESCRIPTION, Default: FALSE}

\item{alpha_per_line}{PARAM_DESCRIPTION, Default: 0.1}

\item{alpha_geom_ribbon}{PARAM_DESCRIPTION, Default: 0.5}

\item{jitter_median_factor}{PARAM_DESCRIPTION, Default: 0}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
