% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparci.R
\name{ggparci}
\alias{ggparci}
\title{Parallel Coordinates Plot with Groups medians and their Confidence Intervals}
\usage{
ggparci(data, columns = 1:ncol(data), groups_column, conf_level = 0.95,
  title = "", groups_in_facets = FALSE, flip_coords = FALSE,
  include_obs_lines = FALSE, alpha_obs_lines = 0.1, alpha_ci_bands = 0.5,
  ...)
}
\arguments{
\item{data}{a data.frame. includes the grouping variable}

\item{columns}{variables to include in the plot, as character vector with columns names or as integer vector with positions. the group variable is ignored, if specified.}

\item{groups_column}{grouping variable name (quoted)}

\item{conf_level}{confidence level for the bands, Default: 0.95}

\item{title}{Plot title, Default: ''}

\item{groups_in_facets}{logical. plot each group in a separate plot/facet? Default: FALSE}

\item{flip_coords}{logical. make the coordinates horizontal? Default: FALSE}

\item{include_obs_lines}{logical. add lines for the rows? Default: FALSE}

\item{alpha_obs_lines}{transparency factor for the lines, Default: 0.1}

\item{alpha_ci_bands}{transparency factor for the confidence bands, Default: 0.5}

\item{...}{additional arguments to the plot}
}
\value{
a ggplot object
}
\description{
The function plots parallel coordinates (gg)plot, in which each line connects the variables medians for each group.
             Confidence "bands" for the medians are added to each line. This allows the assessment of groups (clusters) separations.
             The variables are normalized to [0,1] scale prior to plotting.
}
\examples{

\dontrun{

ggparci(iris, groups_column = "Species")
ggparci(normalize(iris), groups_column = "Species")
ggparci(percentize(iris), groups_column = "Species")

#select only some of the variables
# plot a line for each observation
# and dont plot the CIs
ggparci(percentize(iris),columns = 1:3, groups_column = "Species", include_obs_lines = T, alpha_ci_bands = 0)

# display each group in a different facet
ggparci(data = normalize(iris), groups_column = "Species", groups_in_facets = T)

# flip the plot
ggparci(data = normalize(iris), groups_column = "Species", flip_coords = T)

}
}
